fix(ui): 修正 Fx 响应式网格列数类型

+++ client/lib/ui/widgets/fx_responsive.dart
@@ -38,7 +38,7 @@ class FxResponsiveFormGrid extends StatelessWidget {
             ((availableWidth + horizontalGap) /
                     (effectiveMinWidth + horizontalGap))
                 .floor();
-        columns = columns.clamp(1, maxColumns);
+        columns = columns.clamp(1, maxColumns).toInt();
         final width =
             (availableWidth - horizontalGap * (columns - 1)) / columns;
